Refactored Profile Page Code into three parts.
